requirements.txt
tweepy
wordcloud
matplotlib
numpy
pymongo
pillow